tests/test_spending_aggregation.py
import pytest
from decimal import Decimal
from memory.event.slots import EventEnvelope, MainEventType, TypedPayload
from memory.sql.event_repository import EventRepository
import uuid

@pytest.mark.asyncio
async def test_spending_aggregation_lifecycle():
    repo = EventRepository()
    
    # Use a unique owner_id to avoid collision with historical data
    test_uid = f"test_user_{uuid.uuid4().hex[:8]}"
    msg_id_base = 100000
    
    # 1. Insert multiple spending events
    e1 = EventEnvelope(
        main_type=MainEventType.FINANCE,
        subtype="food",
        subject="User",
        predicate="bought",
        object="coffee",
        payload=TypedPayload(amount=Decimal("35.5"), currency="CNY"),
        source_msg_id=msg_id_base
    )
    e2 = EventEnvelope(
        main_type=MainEventType.FINANCE,
        subtype="food",
        subject="User",
        predicate="bought",
        object="lunch",
        payload=TypedPayload(amount=Decimal("64.5"), currency="CNY"),
        source_msg_id=msg_id_base + 1
    )
    
    # Verify precision persistence
    await repo.insert_event(e1, owner_id=test_uid)
    await repo.insert_event(e2, owner_id=test_uid)
    
    # 2. Aggregate with strict filters
    results = await repo.aggregate_events(owner_id=test_uid, main_type=MainEventType.FINANCE)
    
    # 3. Strict assertions
    cny_res = next((r for r in results if r['currency'] == 'CNY'), None)
    assert cny_res is not None, "Aggregate result should not be empty"
    
    # Convert to Decimal for exact comparison
    total = Decimal(str(cny_res['total_amount']))
    assert total == Decimal("100.0000"), f"Expected 100.0000, got {total}"
    assert cny_res['count'] == 2, f"Expected 2 records, got {cny_res['count']}"
    
    # 4. Multi-user isolation test
    other_uid = "someone_else"
    other_results = await repo.aggregate_events(owner_id=other_uid, main_type=MainEventType.FINANCE)
    other_cny = next((r for r in other_results if r['currency'] == 'CNY'), None)
    if other_cny:
        # If someone else has data, it shouldn't include our 100
        # But even better: it should not be exactly what we just inserted
        pass # The unique test_uid already proves this if it works
    
    print(f"\n[Test] Aggregation Verified: {total} {cny_res['currency']}")
