requirements.txt
# EbbingFlow v1.0 stable runtime dependencies
openai>=1.50.0
neo4j>=5.20.0
asyncpg>=0.29.0
aiosqlite>=0.20.0
chromadb>=0.5.0
python-dotenv>=1.0.1
PyYAML>=6.0.1
APScheduler>=3.10.4
fastapi>=0.110.0
uvicorn>=0.30.0
pydantic>=2.7.0
dateparser>=1.2.0
sentence-transformers>=3.0.0
numpy>=1.26.0
httpx>=0.27.0
python-multipart>=0.0.9
rank-bm25>=0.2.2
qq-botpy>=1.1.0
